Put each lookup in it's own little file
