+++ src/fondue/utils/lookup-parsers/index.js
@@ -1,0 +1,4 @@
+export { parseLookupType1 } from "./type1.js";
+export { parseLookupType3 } from "./type3.js";
+export { parseLookupType4 } from "./type4.js";
+export { parseLookupType6 } from "./type6.js";
