+++ src/fondue/utils/lookup-parsers/type6.js
@@ -1,74 +1,3 @@
-// Single substitution
-export function parseLookupType1(lookup, charFor, charactersFromGlyphs, mergeUniqueCoverage) {
-	const parsedData = { input: [], backtrack: [], lookahead: [], alternateCount: [] };
-
-	lookup.subtableOffsets.forEach((_, i) => {
-		const subtable = lookup.getSubTable(i);
-		const coverage = subtable.getCoverageTable();
-		const results = charactersFromGlyphs(coverage, charFor);
-
-		if (results.length > 0) {
-			parsedData.input = results;
-		}
-	});
-
-	return parsedData;
-}
-
-// Alternate substitution
-export function parseLookupType3(lookup, charFor, charactersFromGlyphs, mergeUniqueCoverage) {
-	const parsedData = { input: [], backtrack: [], lookahead: [], alternateCount: [] };
-
-	lookup.subtableOffsets.forEach((_, i) => {
-		const subtable = lookup.getSubTable(i);
-		const coverage = subtable.getCoverageTable();
-
-		// It's possible to have AlternateSets with different lengths
-		// inside the same lookup (e.g. 10 alternates for "A", 5 for
-		// "B"), so we keep track of the alternateCount per glyph.
-		subtable.alternateSetOffsets.forEach((_, j) => {
-			parsedData.input = charactersFromGlyphs(coverage, charFor, j);
-
-			const altset = subtable.getAlternateSet(j);
-			parsedData.alternateCount.push(altset.alternateGlyphIDs.length);
-		});
-	});
-
-	return parsedData;
-}
-
-// Ligature substitution
-export function parseLookupType4(lookup, charFor, charactersFromGlyphs, mergeUniqueCoverage) {
-	const parsedData = { input: [], backtrack: [], lookahead: [], alternateCount: [] };
-
-	lookup.subtableOffsets.forEach((_, i) => {
-		const subtable = lookup.getSubTable(i);
-		const coverage = subtable.getCoverageTable();
-
-		if (coverage.glyphArray !== undefined) {
-			subtable.ligatureSetOffsets.forEach((_, setIndex) => {
-				const ligatureSet = subtable.getLigatureSet(setIndex);
-
-				ligatureSet.ligatureOffsets.forEach((_, ligIndex) => {
-					const ligatureTable = ligatureSet.getLigature(ligIndex);
-
-					const sequence = [
-						coverage.glyphArray[setIndex],
-						...ligatureTable.componentGlyphIDs,
-					].map(charFor);
-
-					// Only keep sequences with glyphs mapped to letters
-					if (!sequence.includes(undefined)) {
-						parsedData.input.push(sequence.join(""));
-					}
-				});
-			});
-		}
-	});
-
-	return parsedData;
-}
-
 // Chained context substitution
 export function parseLookupType6(lookup, charFor, charactersFromGlyphs, mergeUniqueCoverage) {
 	const parsedData = { input: [], backtrack: [], lookahead: [], alternateCount: [] };
